Refactor to merge wavfile changes to common
Isolated the changes to WavFile and WavOutputStream so as to not have to make changes to other places in the app until it was finished, now it needed to be moved back to common.
