+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/wav/CueChunk.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.jvm.markerapp.audio.wav
+package org.wycliffeassociates.otter.common.wav
 
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
