+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/wav/WavCue.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.jvm.markerapp.audio.wav
+package org.wycliffeassociates.otter.common.wav
 
 data class WavCue(
     val location: Int,
